+++ src/siun/config.py
@@ -6,7 +6,7 @@ from tomllib import TOMLDecodeError
 from tomllib import load as toml_load
 from typing import Any, Self
 
-from pydantic import BaseModel, Field, ValidationError, computed_field, field_validator, model_validator
+from pydantic import BaseModel, ConfigDict, Field, ValidationError, computed_field, field_validator, model_validator
 
 from siun.errors import ConfigError
 from siun.models import (
@@ -20,6 +20,7 @@ from siun.models import (
     V2Threshold,
 )
 from siun.notification import UpdateNotification
+from siun.providers import UPDATE_PROVIDER_REGISTRY, UpdateProvider, UpdateProviderPacman
 from siun.util import get_default_config_dir, get_default_state_path
 
 
@@ -41,10 +42,15 @@ def get_default_criteria() -> list[V2Criterion]:
     ]
 
 
+def get_default_update_provider() -> UpdateProvider:
+    """Get default update provider."""
+    return UpdateProviderPacman()
+
+
 class SiunConfig(BaseModel):
     """Config struct."""
 
-    cmd_available: str = Field(default="pacman -Quq; if [ $? == 1 ]; then :; fi")
+    update_provider: UpdateProvider = Field(default_factory=get_default_update_provider)
     cache_min_age_minutes: int = Field(default=30)
     v2_thresholds: list[V2Threshold] = Field(default_factory=get_default_thresholds)
     v2_criteria: list[V2Criterion] = Field(default_factory=get_default_criteria)
@@ -77,7 +83,8 @@ class SiunConfig(BaseModel):
 
         has_thresholds = "thresholds" in data
         has_criteria = "criteria" in data
-        if not has_thresholds and not has_criteria:
+        has_cmd_available = "cmd_available" in data
+        if not has_thresholds and not has_criteria and not has_cmd_available:
             return data  # pyright: ignore[reportUnknownVariableType]
 
         message_parts = ["Found deprecated config fields: "]
@@ -85,6 +92,8 @@ class SiunConfig(BaseModel):
             message_parts.append("- 'thresholds' have been deprecated in favour of 'v2_thresholds'")
         if has_criteria:
             message_parts.append("- 'criteria' have been deprecated in favour of 'v2_criteria'")
+        if has_cmd_available:
+            message_parts.append("- 'cmd_available' has been deprecated in favour of 'update_provider'")
 
         raise ValueError("\n".join(message_parts))
 
@@ -120,6 +129,14 @@ class SiunConfig(BaseModel):
             raise ValueError(message)
 
         return value
+
+    @field_validator("update_provider")
+    def transform_update_provider(cls, value: UpdateProvider) -> UpdateProvider:
+        """Transform update provider to subclasses of UpdateProvider."""
+        registry = UPDATE_PROVIDER_REGISTRY
+        return registry.get(value.name)(**value.model_dump())
+
+    model_config = ConfigDict(extra="allow")  # pyright: ignore[reportUnannotatedClassAttribute]
 
 
 def _read_config(config_path: Path) -> dict[str, Any]:  # pragma: no cover