+++ src/siun/models.py
@@ -118,3 +118,13 @@ class CriterionCustom(V2Criterion):
 
     name: str = "custom"
     model_config = ConfigDict(extra="allow")  # pyright: ignore[reportUnannotatedClassAttribute]
+
+
+class PackageUpdate(BaseModel):
+    """Struct representing an available update."""
+
+    name: str
+    old_version: str | None = None
+    new_version: str | None = None
+
+    model_config = ConfigDict(extra="allow")  # pyright: ignore[reportUnannotatedClassAttribute]